requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
Django==3.1
django-cors-headers==3.5.0
django-extensions==3.0.5
idna==2.10
mysqlclient==2.0.1
Pillow==7.2.0
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
svgwrite==1.4
Tree==0.2.4
urllib3==1.25.10
